requirements.txt
streamlit
pandas
openpyxl
python-pptx